requirements.txt
flask>=3.0
mistralai>=1.0
python-dotenv>=1.0
